+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -250,9 +250,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-07-22T08:59:58.831Z"
+    created: Optional[str] = "2026-07-22T10:01:41.367Z"
 
-    updated: Optional[str] = "2026-07-22T08:59:58.831Z"
+    updated: Optional[str] = "2026-07-22T10:01:41.367Z"
 
     updated_by_id: OptionalNullable[str] = UNSET