+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsepython.py
@@ -40,9 +40,9 @@ class EvaluatorResponsePython(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-22T08:59:58.831Z"
+    created: Optional[str] = "2026-07-22T10:01:41.367Z"
 
-    updated: Optional[str] = "2026-07-22T08:59:58.831Z"
+    updated: Optional[str] = "2026-07-22T10:01:41.367Z"
 
     updated_by_id: OptionalNullable[str] = UNSET