+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-22T08:59:49.949Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-22T10:01:36.519Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")