+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY4GW4DWAP8BWNQ7R36891TD"
+        "01KY4MCVR0YRSYPZBAERBM3P7N"
     )
     r"""The unique identifier of the data source"""